pyproject.toml
[project]
name = "memory-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "langchain-openai>=0.2.9",
    "loguru>=0.7.2",
    "pymilvus[model]>=2.4.9",
    "python-dotenv>=1.0.1",
    "redis>=5.2.0",
]
